+++ dart/lib/sentry.dart
@@ -37,3 +37,4 @@ export 'src/exception_cause_extractor.dart';
 export 'src/exception_cause.dart';
 // Isolates
 export 'src/sentry_isolate_extension.dart';
+export 'src/sentry_isolate.dart';